% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{add_interaction_score}
\alias{add_interaction_score}
\title{Normalize and integrate interaction intensity of LR pairs}
\usage{
add_interaction_score(EV_spatalk_object = EV.spatalk.results, mc.cores = 30)
}
\arguments{
\item{EV_spatalk_object}{An updated S4 object that has been prepared for EV_SpaTalk analysis.}

\item{mc.cores}{The number of cores to utilize for parallel processing. This should be adjusted according to the capabilities of your computational environment; set to 1 for single-core machines.}
}
\value{
The EV_spatalk_object with added normalized interaction intensity scores for all LR pairs, incorporated into the metadata of the ST Seurat object.
}
\description{
Applies Min-Max normalization to the interaction intensity of each ligand-receptor (LR) pair, scaling the range to 0-1. Integrates the normalized results into the spatial transcriptomics (ST) Seurat object contained within the EV_spatalk_object.
}
\examples{
EV.spatalk.results <- add_interaction_score(EV_spatalk_object = EV.spatalk.results, mc.cores=30)
}
