% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{calculate_interaction_strength}
\alias{calculate_interaction_strength}
\title{Calculate LR pseudo_count}
\usage{
calculate_interaction_strength(L, R, pseudo_count = 0.01)
}
\description{
Calculate LR pseudo_count
}
\keyword{internal}
