% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EV_SpaTalk.obj.R
\name{create_EV.SpaTalk.obj}
\alias{create_EV.SpaTalk.obj}
\title{create EV_SpaTalk S4 object}
\usage{
create_EV.SpaTalk.obj()
}
\description{
create EV_SpaTalk S4 object
}
\examples{
create_EV.SpaTalk.obj()
}
