% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{find.inter.LR}
\alias{find.inter.LR}
\title{Identify overlapping LR interactomes across niches}
\usage{
find.inter.LR(EV_spatalk_object = EV.spatalk.results, mc.cores = 30)
}
\arguments{
\item{EV_spatalk_object}{An updated S4 object that has been prepared for EV_SpaTalk analysis.}

\item{mc.cores}{The number of cores to use for parallel computation. Adjust according to your server's capabilities; use 1 for a single-core PC.}
}
\value{
Returns the EV_spatalk_object updated with identified overlapping LR pairs and their interaction intensities across all examined niches.
}
\description{
Analyzes the frequency of ligand-receptor (LR) pairs enriched in spatially adjacent spots to identify those that are consistently present across different niches. This function facilitates the parallel analysis of LR interactomes and their interaction intensities.
}
\examples{
EV.spatalk.results <- find.inter.LR(EV_spatalk_object=EV.spatalk.results, mc.cores=30)
}
