% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{find_EV_spatalk_LR}
\alias{find_EV_spatalk_LR}
\title{Statistical analysis of LR pairs for EV secretion and spatial distance.}
\usage{
find_EV_spatalk_LR(
  EV_spatalk_object = EV.spatalk.results,
  mc.cores = 30,
  seeds = 1
)
}
\arguments{
\item{EV_spatalk_object}{An updated S4 object that has been prepared for EV_SpaTalk analysis.}

\item{mc.cores}{The number of cores to utilize for parallel processing. This should be adjusted according to the capabilities of your computational environment; set to 1 for single-core machines.}
}
\value{
An updated EV_spatalk_object including the significance analysis of all LR pairs, with additional annotations for their relevance to EV release and spatial distance correlation.
}
\description{
This function employs the Robust Rank Aggregation (RRA) algorithm to assess the enrichment significance of ligand-receptor (LR) pairs. It also identifies LR pairs that significantly correlate with the EV release capacity of the sender cell and the spatial proximity between sender and receiver cell niches.
}
\examples{
EV.spatalk.results <- find_EV_spatalk_LR(EV_spatalk_object=EV.spatalk.results, mc.cores=30, seeds = 1)
}
