% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{FindLR2}
\alias{FindLR2}
\title{Find LR recurrented in sender spot}
\usage{
FindLR2(data_1, data_2 = NULL, datatype, comm_type, database = NULL)
}
\description{
Find LR recurrented in sender spot
}
\keyword{internal}
