% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{GeneToEnrichment}
\alias{GeneToEnrichment}
\title{Modules to cells}
\usage{
GeneToEnrichment(
  srt,
  type = "GO",
  db = NULL,
  method = "rand",
  genes = NULL,
  assay = NULL,
  do.rescale = FALSE,
  min.cells = 0,
  min.genes = 0,
  min.var = 0,
  min.var.rescaled = 0,
  auc_percentile = 0.05,
  db_rand = NULL,
  nrand = 4,
  nbin = 5,
  ...
)
}
\description{
Modules to cells
}
\keyword{internal}
