% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{LR_circos.plot}
\alias{LR_circos.plot}
\title{circos plot of candidate LRs}
\usage{
LR_circos.plot(
  EV_spatalk_object = EV.spatalk.results,
  select.LR.id = select.LR.id,
  specific.LR.id = "CD96_PVR"
)
}
\description{
circos plot of candidate LRs
}
\keyword{internal}
