% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{LR_pie.plot}
\alias{LR_pie.plot}
\title{Pie plot of the statistical results of EV-release related or distance related LRs}
\usage{
LR_pie.plot(EV_spatalk_object = EV.spatalk.results)
}
\description{
Pie plot of the statistical results of EV-release related or distance related LRs
}
