% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{LR_spatial_indensity_plot}
\alias{LR_spatial_indensity_plot}
\title{Spatial indensity_plot}
\usage{
LR_spatial_indensity_plot(EV_spatalk_object, s.LR.pair)
}
\description{
Spatial indensity_plot
}
\keyword{internal}
