% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_public.R
\name{LR_venn.plot}
\alias{LR_venn.plot}
\title{Venn plot of the LRs overlapping}
\usage{
LR_venn.plot(EV_spatalk_object = EV.spatalk.results)
}
\description{
Venn plot of the LRs overlapping
}
\keyword{internal}
