% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{st_deco_anno}
\alias{st_deco_anno}
\title{Decompose cell types in spatial transcriptomics data}
\usage{
st_deco_anno(
  st = st,
  sc = sc,
  EV_spatalk_object = EV_spatalk_object,
  nrand = 2,
  nbin = 5,
  pval_thresh = 0.05,
  mc.cores = 30,
  sc.anno.id = "pop",
  set.seeds = 1,
  select.celltype = select.celltype,
  pred_bin_cutoff = 0.9,
  coef_bin_cutoff = 2
)
}
\arguments{
\item{EV_spatalk_object}{An S4 object reconstructed for EV_SpaTalk. This object must correctly include both spatial transcriptomics (st) and single-cell RNA-seq (sc) data entries.}

\item{sc.anno.id}{A character containing the cell type of the reference single-cell RNA-seq data.}

\item{select.celltype}{A character the all selected cell types for ST spot deconvolution.}
}
\value{
An updated EV_spatalk_object containing the results of the decomposition analysis, including cell type distributions and spatial distances between identified cell types.
}
\description{
Identify the cellular composition for single-cell or spot-based spatial transcriptomics data with non-negative regression.
}
\examples{
EV.spatalk.results <- st_deco_anno(st = st, sc = sc, nrand = 2, EV_spatalk_object=EV_spatalk_object, nbin = 5, pval_thresh=0.05, mc.cores=30, sc.anno.id="pop", set.seeds = 1, select.celltype = select.celltype)
}
