% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{find_niche_LR}
\alias{find_niche_LR}
\title{find near niche from ST data}
\usage{
find_niche_LR(
  EV_spatalk_object = EV.spatalk.results,
  prox = "inverse",
  mc.cores = 30,
  s.cell.type = c("Malignant", "T_cells"),
  comm_list = comm_list,
  datatype = "mean count",
  method = "pseudocount"
)
}
\arguments{
\item{EV_spatalk_object}{the EV_SpaTalk S4 object.}

\item{sc.anno.id}{the character of cell type annotation in paired scRNA-seq data.}
}
\value{
EV_spatalk_object
}
\description{
This function takes two numbers and returns their sum.
}
\examples{
add(1, 1)
}
