% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_main.R
\name{st_deco_anno}
\alias{st_deco_anno}
\title{Make cell type deconvolution from ST data}
\usage{
st_deco_anno(
  st = st,
  sc = sc,
  EV_spatalk_object = EV_spatalk_object,
  nrand = 2,
  nbin = 5,
  pval_thresh = 0.05,
  mc.cores = 30,
  sc.anno.id = "pop",
  set.seeds = 1,
  select.celltype = select.celltype,
  pred_bin_cutoff = 0.9,
  coef_bin_cutoff = 2
)
}
\arguments{
\item{EV_spatalk_object}{the EV_SpaTalk S4 object.}

\item{sc.anno.id}{the character of cell type annotation in paired scRNA-seq data.}
}
\value{
EV_spatalk_object
}
\description{
This function takes two numbers and returns their sum.
}
\examples{
add(1, 1)
}
